cpp05 ex02 fixes

+++ cpp05/ex02/AForm.hpp
@@ -6,7 +6,7 @@
 /*   By: dximenez <dximenez@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/17 18:31:21 by dximenez          #+#    #+#             */
-/*   Updated: 2024/06/18 10:49:45 by dximenez         ###   ########.fr       */
+/*   Updated: 2024/08/08 22:49:32 by dximenez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,6 +14,7 @@
 #define AFORM_HPP
 
 #include <iostream>
+#include <stdlib.h>
 #include "Bureaucrat.hpp"
 
 class Bureaucrat;
@@ -22,8 +23,8 @@ class AForm
 	private:
 		const std::string	_name;
 		bool				_signed;
-		const uint8_t		_gradeToSign;
-		const uint8_t		_gradeToExecute;
+		const int		_gradeToSign;
+		const int		_gradeToExecute;
 	public:
 		AForm(const std::string name, int gradeToSign, int gradeToExecute);				// Default constructor
 		AForm(const AForm &ref);				// Copy constructor
@@ -36,8 +37,8 @@ class AForm
 
 		std::string	getName() const;
 		bool		getSigned() const;
-		uint8_t		getGradeToSign() const;
-		uint8_t		getGradeToExecute() const;
+		int		getGradeToSign() const;
+		int		getGradeToExecute() const;
 
 		class GradeTooHighException : public std::exception
 		{
